requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
pytest==7.4.0 